refactored the TextInput and its controlling state variable to a separate component
