+++ App.js
@@ -2,26 +2,18 @@ import { StatusBar } from "expo-status-bar";
 import { StyleSheet, Text, TextInput, View } from "react-native";
 import Header from "./components/Header";
 import { useState } from "react";
+import Input from "./components/Input";
 
 export default function App() {
   const appName = "my first app";
-  const [text, setText] = useState("");
+  // const [text, setText] = useState("");
 
-  // callback handler
-  function changeTextHandler(changedText) {
-    setText(changedText);
-  }
   return (
     <View style={styles.container}>
       <StatusBar style="auto" />
       <Header name={appName} version={2} />
-      <TextInput
-        placeholder="Type something"
-        style={styles.input}
-        value={text}
-        onChangeText={changeTextHandler}
-      />
-      <Text>{text}</Text>
+      <Input />
+      {/*<Text>{text}</Text>*/}
     </View>
   );
 }
@@ -32,10 +24,5 @@ const styles = StyleSheet.create({
     backgroundColor: "#fff",
     alignItems: "center",
     justifyContent: "center",
-  },
-  input: {
-    borderBottomWidth: 2,
-    borderBottomColor: "purple",
-    width: "50%",
   },
 });
